# api/api.yml
openapi: 3.0.2
info:
  version: 1.0.0
  title: Issue Tracker API
  description: >
    REST API for the **Issue Tracker** (Taiga clone) built for the ASW course at UPC.
    The web application is available at https://projecte-asw1.onrender.com.


    **Authentication**: All endpoints require an API key in the `Authorization` header.
    Find your key on your profile page in the web application.
servers:
  - url: 'https://projecte-asw1.onrender.com/api'
security:
  - ApiKeyAuth: []
paths:
...
  /issues/{issue_id}/:
    get:
      tags: [issues]
      summary: Get detail of an issue
      operationId: getIssue
      parameters:
        - name: issue_id
          in: path
          required: true
          schema:
            type: integer
          example: 42
      responses:
        "200":
          description: Issue detail with nested data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [issues]
      summary: Delete an issue
      operationId: deleteIssue
      parameters:
        - name: issue_id
          in: path
          required: true
          schema:
            type: integer
          example: 42
      responses:
        "204":
          description: Issue deleted successfully
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
    put:
      tags: [issues]
      summary: Update an issue (only creator)
      operationId: updateIssue
      parameters:
        - name: issue_id
          in: path
          required: true
          schema:
            type: integer
          example: 42
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueUpdateInput'
      responses:
        "200":
          description: Issue updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
...
  /issues/{issue_id}/deadline/:
    post:
      tags: [issues]
      summary: Set the deadline for an issue
      operationId: setIssueDeadline
      parameters:
        - name: issue_id
          in: path
          required: true
          schema:
            type: integer
          example: 42
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deadline:
                  type: string
                  format: date
                  description: "The deadline date (YYYY-MM-DD)."
                  example: '2026-06-01'
      responses:
        "200":
          description: Updated issue details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [issues]
      summary: Remove the deadline for an issue
      operationId: deleteIssueDeadline
      parameters:
        - name: issue_id
          in: path
          required: true
          schema:
            type: integer
          example: 42
      responses:
        "200":
          description: Updated issue details (without deadline)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: >
        Your personal API key. Find it on your profile page in the web application.
        Send the raw 64-character hex value — no prefix required.
  responses:
    Unauthorized:
      description: 'Error: Unauthorized — API key missing or invalid'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
          examples:
            error401:
              value:
                message: 'Invalid API key.'
    Forbidden:
      description: 'Error: Forbidden — you do not have permission for this action'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
          examples:
            error403:
              value:
                message: 'You do not have permission to perform this action.'
    NotFound:
      description: 'Error: Not Found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
          examples:
            error404:
              value:
                message: "No resource with the given id found."
    BadRequest:
      description: 'Error: Bad Request'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
          examples:
            error400:
              value:
                message: 'subject: This field is required.'
  schemas:
    ErrorResult:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: 'Invalid API key.'
...
    IssueStatus:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'New'
        slug:
          type: string
          example: 'new'
        color:
          type: string
          example: '#83eede'
        is_closed:
          type: boolean
          example: false
    IssueUser:
      type: object
      properties:
        username:
          type: string
          example: john
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
    IssueTag:
      type: object
      properties:
        name:
          type: string
          example: 'backend'
        slug:
          type: string
          example: 'backend'
        color:
          type: string
          example: '#70728f'
    IssueList:
      type: object
      properties:
        id:
          type: integer
          example: 42
        subject:
          type: string
          example: 'Login button not working on mobile'
        status:
          $ref: '#/components/schemas/IssueStatus'
        issue_type:
          type: string
          example: 'bug'
        severity:
          type: string
          example: 'critical'
        priority:
          type: string
          example: 'high'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/IssueTag'
        deadline:
          type: string
          format: date
          nullable: true
          example: '2026-06-01'
        deadline_status:
          type: string
          nullable: true
          enum: [red, orange, green]
          example: 'orange'
        assigned_to:
          allOf:
           - $ref: '#/components/schemas/IssueUser'
          nullable: true
        created_by:
          $ref: '#/components/schemas/IssueUser'
        created_at:
          type: string
          format: date-time
          example: '2026-05-01T10:00:00Z'
        modified_at:
          type: string
          format: date-time
          example: '2026-05-03T14:30:00Z'
    IssueDetail:
      allOf:
        - $ref: '#/components/schemas/IssueList'
        - type: object
          properties:
            description:
              type: string
              example: 'When tapping the login button on iOS Safari, nothing happens.'
...
    IssueUpdateInput:
      type: object
      properties:
        subject:
          type: string
          example: 'Updated subject'
        description:
          type: string
          example: 'Updated description'
        status_id:
          type: integer
          example: 2
        issue_type:
          type: string
          example: 'bug'
        severity:
          type: string
          example: 'Normal'
        priority:
          type: string
          example: 'Normal'
        tags:
          type: array
          items:
            type: string
          example: []
        assigned_to_id:
          type: integer
          nullable: true
          example: 5
        deadline:
          type: string
          format: date
          nullable: true
          example: '2026-06-10'